// types.ts
export enum UserRole {
  ADMIN = 'ADMIN',
  TEACHER = 'TEACHER',
  PARENT = 'PARENT',
  BOARD_MEMBER = 'BOARD_MEMBER',
}

export enum SchoolStatus {
    ACTIVE = 'ACTIVE',
    CAUTION = 'CAUTION',
    SUSPENDED = 'SUSPENDED'
}

export interface School {
    id: string;
    name: string;
    address: string;
    motto?: string;
    domain: string;
    badgeUrl?: string | null;
    status: SchoolStatus;
    cautionMessage: string;
    activeSubjectIds: string[];
}


export interface User {
  id: string;
  username: string;
  password?: string; // Password stored here for mock auth
  roles: UserRole[];
  name: string;
  phoneNumber?: string; // For SMS notifications
  studentId?: string; // For parents, links to their child
  teacherId?: string; // For teachers
  profileImageUrl?: string; // For teacher profile images
  signatureImageUrl?: string; // For teacher/admin signatures
  classSubjectAssignments?: { classId: string; subjectIds: string[] }[];
  schoolId?: string; // To associate user with a school. Optional for superadmin.
  isBoardMember?: boolean;
}
// ...
export interface SchoolClass {
  id: string;
  name: string; // e.g., "Form 1", "Form 2"
}

export interface Subject {
  id: string;
  name: string;
  group?: number; // To categorize subjects (1=Compulsory, 2=Sciences, etc.)
}
// ...
export enum AssignmentType {
    ASSIGNMENT = 'ASSIGNMENT',
    MATERIAL = 'MATERIAL',
}

export interface Assignment {
  id: string;
  title: string;
  content: string; // Markdown content
  teacherId: string;
  classId: string;
  subjectId: string;
  dueDate?: string; // Optional due date for assignments
  createdAt: string;
  type: AssignmentType;
  fileUrl?: string;
  fileName?: string;
  fileType?: string;
  schoolId: string;
}